+++ app/controllers/devise/sessions_controller.rb
@@ -1,6 +1,7 @@
 class Devise::SessionsController < DeviseController
   prepend_before_filter :require_no_authentication, only: [ :new, :create ]
   prepend_before_filter :allow_params_authentication!, only: :create
+  prepend_before_filter :verify_signed_out_user, only: :destroy
   prepend_before_filter only: [ :create, :destroy ] { request.env["devise.skip_timeout"] = true }
 
   # GET /resource/sign_in
@@ -49,5 +50,28 @@ class Devise::SessionsController < DeviseController
 
   def auth_options
     { scope: resource_name, recall: "#{controller_path}#new" }
+  end
+
+  private
+
+  # Check if there is no logged user before doing the logout.
+  #
+  # If there is no logged in user, it will set the flash message and redirect
+  # to the after_sign_out path.
+  def verify_signed_out_user
+    if user_already_signed_out?
+      set_flash_message :notice, :already_signed_out if is_flashing_format?
+
+      respond_to do |format|
+        format.all { head :no_content }
+        format.any(*navigational_formats) { redirect_to after_sign_out_path_for(resource_name) }
+      end
+    end
+  end
+
+  def user_already_signed_out?
+    users = Devise.mappings.keys.map { |s| warden.user(scope: s, run_callbacks: false) }
+
+    users.all?(&:blank?)
   end
 end